backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
Cython==3.0.6
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
MouseInfo==0.1.3
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.0
pyobjc-framework-ApplicationServices==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pytweening==1.0.7
rubicon-objc==0.4.7
screeninfo==0.8.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tk==0.1.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
